requirements.txt
pycrypto==2.6.1
pysha3==1.0b1
